+++ README.md
@@ -1,67 +1,63 @@
-# README
+**ER Diagram**
 
-## FURIMA ER Diagram
-### Entities and Attributes
+**Users**
 
-#### USERS
-- **Relationships**:
-  - Has many ITEMS (One to Many)
-  - Has many ORDERS (One to Many)
-  - USERS can exist without ITEMS
-  - USERS can exist without ORDERS
-- **Attributes**:
-  - User_ID (Primary Key)
-  - Username
-  - Password
-  - Email
-  - Address
+| Column             | Type      | Options                   |
+| ------------------ | --------- | ------------------------- |
+| nickname           | string    | null: false               |
+| email              | string    | null: false, unique: true |
+| encrypted_password | string    | null: false               |
+| last_name          | string    | null: false               |
+| first_name         | string    | null: false               |
+| last_name_kana     | string    | null: false               |
+| first_name_kana    | string    | null: false               |
+| birth_date         | date      | null: false               |
 
-#### ITEMS
-- **Relationships**:
-  - Belongs to USERS (Many to One)
-  - Has many ORDERS (One to Many)
-  - ITEMS cannot exist without USERS
-  - ITEMS can exist without ORDERS
-- **Attributes**:
-  - Item_ID (Primary Key)
-  - Name
-  - Category
-  - Price
-  - User_ID (Foreign Key)
+**Association:**
+- has_many :items
+- has_many :orders
 
-#### ORDERS
-- **Relationships**:
-  - Belongs to one USERS (Many to One)
-  - Belongs to one ITEMS (Many to One)
-  - ORDERS cannot exist without USERS
-  - ORDERS cannot exist without ITEMS
-- **Attributes**:
-  - Order_ID (Primary Key)
-  - User_ID (Foreign Key)
-  - Item_ID (Foreign Key)
-  - Shipping_Address (Each Order can have only one Shipping Address, but multiple Orders can share the same Shipping Address)
+**Items**
 
+| Column             | Type      | Options                   |
+| ------------------ | --------- | ------------------------- |
+| user               | references| null: false, foreign_key: true|
+| name               | string    | null: false               |
+| description        | text      | null: false               |
+| category_id        | integer   | null: false               |
+| condition_id       | integer   | null: false               |
+| delivery_fee_id    | integer   | null: false               |
+| prefecture_id      | integer   | null: false               |
+| shipping_day_id    | integer   | null: false               |
+| price              | integer   | null: false               |
 
+**Association:**
+- belongs_to :user
+- has_one :order
 
-This README would normally document whatever steps are necessary to get the
-application up and running.
+**Orders**
 
-Things you may want to cover:
+| Column             | Type      | Options                   |
+| ------------------ | --------- | ------------------------- |
+| user               | references| null: false, foreign_key: true|
+| item               | references| null: false, foreign_key: true|
 
-* Ruby version
+**Association:**
+- belongs_to :user
+- belongs_to :item
+- has_one :shipping_address
 
-* System dependencies
+**ShippingAddresses**
 
-* Configuration
+| Column             | Type      | Options                   |
+| ------------------ | --------- | ------------------------- |
+| order              | references| null: false, foreign_key: true|
+| postal_code        | string    | null: false               |
+| prefecture_id      | integer   | null: false               |
+| city               | string    | null: false               |
+| street             | string    | null: false               |
+| building_name      | string    |                           |
+| phone_number       | string    | null: false               |
 
-* Database creation
-
-* Database initialization
-
-* How to run the test suite
-
-* Services (job queues, cache servers, search engines, etc.)
-
-* Deployment instructions
-
-* ...
+**Association:**
+- belongs_to :order
