+++ app/controllers/application_controller.rb
@@ -1,18 +1,34 @@
 class ApplicationController < ActionController::Base
-  before_action :basic_auth
+# USED INSTEAD OF REGISTRATIONS_CONTROLLERS 
+# or SESSIONS_CONTROLLERS
 
-  private
-  def basic_auth
-    authenticate_or_request_with_http_basic do |username, password|
-      username == ENV["BASIC_AUTH_USER"] && password == ENV["BASIC_AUTH_PASSWORD"]
+  before_action :configure_permitted_parameters, if: :devise_controller?
+#  before_action :configure_sign_up_params, only: [:create]
+#  before_action :basic_auth
+
+    def create
+      build_birth_date
+      super
+    end
+  
+    def configure_sign_up_params
+      devise_parameter_sanitizer.permit(:sign_up, keys: [:nickname, :email, :password, :password_confirmation, :last_name, :first_name, :last_name_kana, :first_name_kana, "birth_date(1i)", "birth_date(2i)", "birth_date(3i)"])
+    end
+
+    private
+  
+    def configure_permitted_parameters
+      devise_parameter_sanitizer.permit(:sign_up, keys: [:nickname, :email, :password, :password_confirmation, :last_name, :first_name, :last_name_kana, :first_name_kana, "birth_date(1i)", "birth_date(2i)", "birth_date(3i)"])
+    end
+  
+    def build_birth_date
+      date_params = params[:user].slice("birth_date(1i)", "birth_date(2i)", "birth_date(3i)")
+      begin
+        params[:user][:birth_date] = Date.new(date_params["birth_date(1i)"].to_i, date_params["birth_date(2i)"].to_i, date_params["birth_date(3i)"].to_i)
+      rescue ArgumentError
+        params[:user][:birth_date] = nil
+        @user = User.new(params[:user])  # Initialize a User object to add errors to, if it's not already done
+        @user.errors.add(:birth_date, "is not a valid date")
+      end
     end
   end
-
-  def after_sign_in_path_for(resource)
-    # Customize the path to redirect users after login
-    # For example, redirect to the user's profile page:
-    root_path # or some other path like dashboard_path
-  end
-
-
-end
