+++ app/controllers/users_controller.rb
@@ -1,5 +1,6 @@
 class UsersController < ApplicationController
   def index
+#    user = User.find(params[:id])
     redirect_to root_path
   end
 end
