+++ app/models/ShippingAddress.rb
@@ -1,7 +1,7 @@
 class ShippingAddress < ApplicationRecord
   belongs_to :order
   extend ActiveHash::Associations::ActiveRecordExtensions
-  belongs_to :prefecture
+  belongs_to :prefecture_id
 
   # Validations
   validates :postal_code, :city, :street, :phone_number, presence: true
