+++ config/routes.rb
@@ -1,8 +1,6 @@
 Rails.application.routes.draw do
-  devise_for :users, controllers: {
-    registrations: 'users/registrations',
-    sessions: 'users/sessions'
-  }
+  devise_for :users
+  
   get 'profile', to: 'profiles#show', as: 'user_profile'
 
   # Root route
